requirements.txt
web3
requests
python-decouple
python-telegram-bot==13.15  # Versión estable
flask                   # Opcional: si usas servidor de autorización web
gunicorn                # Para despliegue en VPS